fix: Dockerfile
Hugo extended version requires dynamic libc to run, the Scratch image does not have that so it cause error.
I previous tested the Dockerfile on my Linux machine which build and run the image successfully, because I have a pre-built public folder that did not get excluded while testing the Dockerfile.
So I also add an .dockerignore file to prevent such confusion in the future.